feat: export shows (WIP)
