+++ cmd/import.go
@@ -3,20 +3,12 @@ package cmd
 import (
 	"context"
 	"fmt"
-	"log"
-	"net/http"
-	"os"
 
-	"github.com/elliotwms/opml-to-spotify/pkg/pkce"
 	"github.com/gilliek/go-opml/opml"
-	"github.com/google/uuid"
 	"github.com/spf13/cobra"
 	"github.com/zmb3/spotify/v2"
-	"github.com/zmb3/spotify/v2/auth"
-	"golang.org/x/oauth2"
 )
 
-const flagClientID = "client-id"
 const flagDryRun = "dry-run"
 
 const maxSaveShowsBatchSize = 50
@@ -43,9 +35,7 @@ var (
 func init() {
 	rootCmd.AddCommand(importCmd)
 
-	importCmd.Flags().StringP(flagClientID, "c", "", "Spotify application Client ID")
 	importCmd.Flags().BoolP(flagDryRun, "d", false, "Read-only, will not update your subscriptions")
-
 }
 
 func run(cmd *cobra.Command, args []string) {
@@ -60,7 +50,7 @@ func run(cmd *cobra.Command, args []string) {
 		panic(err)
 	}
 
-	client := spotify.New(login(cmd))
+	client := login(cmd)
 
 	cmd.Printf("Searching for %d shows", len(outlines))
 
@@ -90,21 +80,6 @@ func run(cmd *cobra.Command, args []string) {
 	}
 }
 
-// setClientID sets the clientID var depending on priority from:
-// * initial value (empty but can be set by ldflags)
-// * SPOTIFY_ID environment variable
-// * client-id flag
-func setClientID(cmd *cobra.Command) {
-	if s := os.Getenv("SPOTIFY_ID"); s != "" {
-		clientID = s
-	}
-
-	// allow overriding of client ID via flag
-	if s := cmd.Flag(flagClientID).Value.String(); s != "" {
-		clientID = s
-	}
-}
-
 // getOutlines gets the outlines from a specified OPML file
 func getOutlines(filename string) ([]opml.Outline, error) {
 	f, err := opml.NewOPMLFromFile(filename)
@@ -113,53 +88,6 @@ func getOutlines(filename string) ([]opml.Outline, error) {
 	}
 	outlines := f.Outlines()
 	return outlines, nil
-}
-
-// login logs the user into the application via OAuth authorization code flow
-// https://developer.spotify.com/documentation/general/guides/authorization/code-flow/
-// The user will be presented with a Spotify Login URL via the terminal which they should visit, then be redirected to a
-// locally hosted http server which captures the auth code and performs token exchange
-func login(cmd *cobra.Command) *http.Client {
-	verifier := pkce.NewVerifier(pkce.LenMax)
-
-	s := http.Server{
-		Addr: "localhost:8080",
-	}
-
-	state := uuid.New().String()
-
-	auth := spotifyauth.New(
-		spotifyauth.WithClientID(clientID),
-		spotifyauth.WithScopes(spotifyauth.ScopeUserLibraryModify),
-		spotifyauth.WithRedirectURL("http://localhost:8080/callback"),
-	)
-
-	tokenChan := make(chan *oauth2.Token, 1)
-	http.HandleFunc("/callback", func(writer http.ResponseWriter, request *http.Request) {
-		token, err := auth.Token(context.Background(), state, request, verifier.Params()...)
-
-		if err != nil {
-			log.Fatalf(err.Error())
-		}
-
-		tokenChan <- token
-
-		_, _ = writer.Write([]byte("Token retrieved successfully. Please return to your terminal"))
-	})
-
-	go func() {
-		err := s.ListenAndServe()
-		if err != nil {
-			panic(err)
-		}
-	}()
-
-	challenge := verifier.Challenge()
-	url := auth.AuthURL(state, challenge.Params()...)
-
-	cmd.Printf("Visit this URL in your browser to log in: %s\n", url)
-
-	return auth.Client(context.Background(), <-tokenChan)
 }
 
 // searchSpotifyForOutlines searches the Spotify API for each of the shows specified in the opml outlines by name,
